rename deleteConnection to deleteConnectionByNodeId and fix it. add deleteNodeByUserId.

+++ backend/Services/DataService.cs
@@ -53,12 +53,23 @@ namespace family_tree_API.Services
             return false; ;
         }
 
-        void deleteConnection(String nodeId)
+        void deleteConnectionByNodeId(String nodeId)
         {
             List<Connection> connections = _context.Connections.Where(c=>(c.To.ToString()==nodeId || c.From.ToString() == nodeId)).ToList();
-            _context.Connections.RemoveRange(connections);
+            if (connections.Count > 0)
+            {
+                _context.Connections.RemoveRange(connections);
+            }
         }
 
-
+        void deleteNodeByUserId(String userId)
+        {
+            Node node = _context.Nodes.Where(n => n.FamilyMember.ToString() == userId).FirstOrDefault();
+            if (node != null)
+            {
+                deleteConnectionByNodeId(node.Id.ToString());
+                _context.Nodes.Remove(node);
+            }
+        }
     }
 }
